Move name resolving into copy composer
