+++ internal/databases/postgres/copy_tar_ball_composer.go
@@ -49,6 +49,7 @@ type CopyTarBallComposer struct {
 	fileInfo               map[string]*fileInfo
 	headerInfos            map[string]*headerInfo
 	copiedTars             map[string]bool
+	nameResolver           TarCopiesNameResolver
 }
 
 type CopyTarBallComposerMaker struct {
@@ -163,7 +164,7 @@ func (c *CopyTarBallComposer) copyTar(tarName string) {
 		c.prevBackup.Name+internal.TarPartitionFolderName+tarName,
 		c.newBackupName+internal.TarPartitionFolderName+tarName)
 	c.copiedTars[tarName] = true
-	c.tarBallQueue.TarBallMaker.AddCopiedTarName(tarName)
+	c.nameResolver.AddCopiedTar(tarName)
 	for _, fileName := range c.prevTarFileSets[tarName] {
 		if file, exists := c.fileInfo[fileName]; exists {
 			file.status = processed
@@ -179,7 +180,7 @@ func (c *CopyTarBallComposer) copyTar(tarName string) {
 
 func (c *CopyTarBallComposer) getTarBall() internal.TarBall {
 	tarBall := c.tarBallQueue.Deque()
-	tarBall.SetUp(c.crypter)
+	tarBall.SetUp(c.crypter, c.nameResolver.ResolveId(tarBall.Id(), tarBall.FileExtension()))
 	return tarBall
 }
