+++ internal/databases/postgres/tar_copies_name_resolver.go
@@ -1,4 +1,4 @@
-package internal
+package postgres
 
 import (
 	"fmt"
@@ -22,14 +22,7 @@ func (r *TarCopiesNameResolver) AddCopiedTar(tarName string) {
 	r.copiedTarNames[tarName] = true
 }
 
-func (r *TarCopiesNameResolver) ResolveByName(name string) string {
-	if copied := r.copiedTarNames[name];copied {
-		panic("Tar with this name already exists")
-	}
-	return name
-}
-
-func (r *TarCopiesNameResolver) ResolveByPart(part int, fileExtention string) string {
+func (r *TarCopiesNameResolver) ResolveId(part int, fileExtention string) string {
 	if id, exists := r.tarPartIds[part];exists {
 		return r.defaultName(id, fileExtention)
 	}
