+++ internal/storage_tar_ball_maker.go
@@ -5,11 +5,10 @@ type StorageTarBallMaker struct {
 	partCount       int
 	backupName      string
 	uploader        *Uploader
-	tarNameResolver *TarCopiesNameResolver
 }
 
 func NewStorageTarBallMaker(backupName string, uploader *Uploader) *StorageTarBallMaker {
-	return &StorageTarBallMaker{0, backupName, uploader, NewTarCopiesNameResolver()}
+	return &StorageTarBallMaker{0, backupName, uploader}
 }
 
 // Make returns a tarball with required storage fields.
@@ -25,10 +24,5 @@ func (tarBallMaker *StorageTarBallMaker) Make(dedicatedUploader bool) TarBall {
 		backupName:   tarBallMaker.backupName,
 		uploader:     uploader,
 		partSize:     &size,
-		resolver:     tarBallMaker.tarNameResolver,
 	}
-}
-
-func (tarBallMaker *StorageTarBallMaker) AddCopiedTarName(tarName string) {
-	tarBallMaker.tarNameResolver.copiedTarNames[tarName] = true
 }
