+++ internal/tar_ball_maker.go
@@ -4,5 +4,4 @@ package internal
 // flexible creation of different TarBalls.
 type TarBallMaker interface {
 	Make(dedicatedUploader bool) TarBall
-	AddCopiedTarName(tarName string)
 }
